Fix data types conversion.

+++ src/main/scala/encry/contracts/EncryStateView.scala
@@ -10,7 +10,7 @@ case class EncryStateView(height: Height, lastBlockTimestamp: Long, stateDigest:
   override val tpe: Types.Product = Types.EncryState
 
   override def asVal: PValue = PValue(PObject(Map(
-    "height" -> PValue(height, Types.PInt),
+    "height" -> PValue(height.toLong, Types.PInt),
     "lastBlockTimestamp" -> PValue(lastBlockTimestamp, Types.PInt),
     "stateDigest" -> PValue(stateDigest, Types.PCollection.ofByte)
   ), tpe), tpe)
